+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/update/mappers/UpdateProductUiMappers.kt
@@ -10,7 +10,7 @@ fun UpdateProductUiModel.toDomain(): ProductDomainModel {
     return ProductDomainModel(
         id = this.id,
         name = this.name,
-        price = this.price?.toFloat() ?: 0f,
+        price = this.price?.toDouble() ?: 0.0,
         currency = this.currency,
         unit = this.unit,
         details = this.details,
